Add a way to launch slime with a specific communication style
Adds an argument to slime-init-command to specify additional arguments
to swank:start-server.

Also add slime-style to launching slime with a specific style. This
command I think needs some love to make it fit with the rest of slime
but it felt rude not to include something that got across the idea of
what I need.